[ agda/agda#4560 ] keyword 'pattern' needed now
The fix of agda/agda#4560 in PR agda/agda#4611 requires the record
directive 'pattern' if you want to pattern match on a record constructor
that does not have eta-equality.  Eta-equality is automatically turned
off for some recursive records, such as the present one.

Alternatively, we could add 'eta-equality' instead of 'pattern'.